+++ src/latents/__init__.py
@@ -4,7 +4,6 @@ import contextlib
 from importlib.metadata import PackageNotFoundError, version
 
 from latents import (
-    base,
     gfa,
     mdlag,
     observation_model,
@@ -16,7 +15,6 @@ with contextlib.suppress(PackageNotFoundError):
     __version__ = version("latents")
 
 __all__ = [
-    "base",
     "gfa",
     "mdlag",
     "observation_model",